Fixes to directory pathing - the specified path should now be relative to the NSDocumentDirectory root for the iOS/Mac User directory.
I should update this to handle saving to a different directories than
subdirectories of the NSDocumentDirectory path. I'm not sure if I should
just revert it to the prior behaviour (meaning users will have to create
and specify an absolute path themselves in order to correctly persist
the file). Either way, having it create intermediate directories fixes
the fact that it was silently failing to create local files.