Simplifying the iOS vs OSX code.
There were a couple of methods that were exact duplicates for iOS and OSX except the input and return image types. This change adds an AFImage macro that will be replaced with UIImage when compiling for iOS and NSImage for OSX. This commit also switches to the 'nicer looking' macro TARGET_OS_IPHONE for the platform #ifs.