Initial Check In
Initial Check In.
Slight modification had to be made in order to support this in the HTTP
operation.  I exposed the HTTPError, and added logic to the completion
block to fail with error if an error is present.  In the past, if an
operation was cancelled it simply returned without calling the fail
block.  The combination of exposing the error and adding the check
allows the operation to be canceled, and the error reported to the user.