added will-change state notifications to AFURLConnectionOperation
AFHttpRequestOperation completion blocks are currently called after the operation changes state, so if you are watching the state of the operations in the queue you'll actually see operations finish before their completion blocks have been called. This notification gives you a hook to do something with the operation immediately before the state changes to start or finish.