Automatically dispatch callbacks to the calling queue
If the programmer hasn't already specified a specific queue
for the operation to call back on, assume that the queue calling
the `setCompletionBlockWithSuccess:failure:` selector is the
intended callback queue.

This allows `AFHTTPRequestOperation`, its subclasses, and the
get/post/put/delete/patch methods on `AFHTTPClient`, to be used
safely from long-lived background queues, such as an `NSOperation`
to perform a multi-request action.