add a Xcode project with a static library target
This commits adds a single target project which
can build a static library containing the AFNetworking code.
The library can be built for macos & iphoneos (device + simulator).
Most of the configuration is based on xcconfig files for easier
maintenance (especially diffing and conflict resulution).

The static library project was also added to the AFNetworking
workspace, but the example targets were not touched.