AFHTTPRequestProcessingOperation: Refactor processing support code
The goal of these changes it to simplify the behavior of processing subclasses, avoid code duplication, and make it easier to implement a processing subclass.

All processing is now done in `-processResponse` which is invoked from the completion block. The processing code shares a single dispatch queue, though the queue can be overridden for a given operation.

In the case of AFImageRequestOperation, the processing block is now invoked by the core logic instead of the block in the factory method. The block is now a property and for naming clarity is called filterBlock instead of imageProcessingBlock.