Reachability Manager of AFHTTPRequestOperationManager and AFHTTPSessionManager is now lazily initialized. If it was not used then not created.
I made this for fixing the startMonitoring call during startup, which took 20+ seconds when the network was bad, making the system to kill the app because it took too long to launch in version 2.0.3. Now the status checking is already deferred in the head version, I still would like to have them lazily initialized.