[Issue #1735] Fixing dead store issues in AFSecurityPolicy.m
Build and analyze yields the following issues when assertions are enabled:

```
AFNetworking/AFSecurityPolicy.m:123:18: Value stored to 'status' during its initialization is never read
AFNetworking/AFSecurityPolicy.m:127:9: Value stored to 'status' is never read
```

This patch mimics the pattern established previously in this file on line 89. It adds a check for whether NS_BLOCK_ASSERTIONS is defined to determine whether to use the status variable.