Use SecTrust return status code at runtime
The `SecTrust` functions in use within AFSecurityPolicy.h only make use of the return code when `NS_BLOCK_ASSERTIONS` is not defined. The use of NS_BLOCK_ASSERTIONS to silence static analyzer warnings about unused values is convenient, however the larger issue is the failure of `AFPublicKeyForCertificate`, `AFPublicKeyTrustChainForServerTrust` and `AFServerTrustIsValid` to detect a runtime status code that is not `errSecSuccess` and to return early.