Call super copyWithZone: in subclasses.
By not calling the superclass' implementation, we end up not actually
copying over all the properties from the superclass. In this case, if we
call -copy on an instance of an AFJSONRequestSerializer, the headers for
the serializer are not copied over (as well as anything else that was
changed in the AFHTTPRequestSerializer implementation).