Update UIImageView+AFNetworking.m
Changed the order to first cache the responseObject, and then call the block.
This way the user can modify the downloaded image, and use the modified one as the cached one (with [[UIImageView sharedImageCache] cacheImage:newImage forRequest:imageRequest] )

i.e. if the user wants to downscale an image, he can cache the downscaled to save memory, and allow more objects in cache.