Update AFURLResponseSerialization.m
An http call may return a response with no 'Content-Type' header (it is not mandatory following HTTP spec). If this happens, with the current implementation, there is no way for a user to specify that an empty 'Content-Type' is acceptable if the user has defined a set of acceptable content types.

To fix this, it is enough to use the null object instead of nil when checking if the current content type exist within the acceptable content types. If a user wants to allow an empty content type, he only needs to add [NSNull null] to the acceptable content types list.