Added rss+xml content type to AFURLResponseSerialization.m..
Added @"application/rss+xml" to NSSet of acceptableContentTypes for AFXMLDocumentResponseSerializer and AFXMLParserResponseSerializer.

One specific RSS feed I wanted to parse fails to create a NSXMLParse using AFXMLParseResponseSerializer because they tag their data with the "rss+xml" content type:

http://www.fda.gov/AboutFDA/ContactFDA/StayInformed/RSSFeeds/TDS/rss.xml

If you feed the data from the site directly to a NSXMLParser, it has no problem parsing it. But, AFXMLParserResponseSerializer was failing to return a NSXMLParser because the "rss+xml" data type wasn't recognized.