Fix AFNetworkActivityIndicatorManager for iOS 7...dubiously
This swizzle used to assume that all `NSURLSessionTask` instances had
a superclass named `NSURLSessionTask`. That is not true on iOS
7 devices. Therefore, we need to somehow swizzle the the superclass,
which is `__NSCFLocalSessionTask` on iOS 7. This is a way to do it
without explicitly mentioning or calling any private API.

However, it does end up adding and swizzling methods on
`__NSCFLocalSessionTask`. Is this legal? I cannot say for sure.