Move details of XML and JSON parsing into separate virtual methods
It is not possible today to inherit from AFJSONResponseSerializer and to
easily replace the JSON parser with something else (i.e. SwiftyJSON). This
PR moves the details of XML and JSON parsing (i.e. the exact class) into
separate methods that can be overloaded so that the actual JSON parsing
can be done with another library.