Remove unnecessary certificate validation
When full certificate chain validation was removed, the code to validate the
complete chain wasn't removed but slightly modified so that only 1 certificate
had to be present. This validation is redundant, because the pinned
certificates are already set as anchors for the validation made in the previous
steps. In the case of certificates with multiple certification paths, this
sometimes caused validations to fail which then required adding extra
certificates to the `pinnedCertificates`.