Fixed status code 205 handling
If a server returned a 205 with no response content type and no response data, the response serializer would fail. This patch allows response serialization to continue without error if no response content type header is defined, and there are zero bytes of response of data.

If the server returns a response content type header OR response data, the serializar will respect the old serialization logic. This means that if a response content type is listed in the header, we are expecting some type of data response.